+++ test-app/tests/cucumber/features/step_definitions/my_steps.js
@@ -1,0 +1,32 @@
+(function () {
+
+  'use strict';
+
+  module.exports = function () {
+
+    // You can use normal require here, cucumber is NOT run in a Meteor context (by design)
+    var url = require('url');
+
+    this.Given(/^I have authored the site title as "([^"]*)"$/, function (title, callback) {
+      // this.ddp is a connection to the mirror available to you in all steps
+      this.ddp.call('updateTitle', [title], callback);
+    });
+
+
+    this.When(/^I navigate to "([^"]*)"$/, function (relativePath, callback) {
+      // this.browser is a pre-configured WebdriverIO + PhantomJS instance
+      this.browser.
+        url(url.resolve(process.env.HOST, relativePath)). // process.env.HOST always points to the mirror
+        call(callback);
+    });
+
+    this.Then(/^I should see the heading "([^"]*)"$/, function (expectedTitle, callback) {
+      // you can use chai-as-promised in step definitions also
+      this.browser.
+        waitForVisible('h1'). // WebdriverIO chain-able promise magic
+        getText('h1').should.become(expectedTitle).and.notify(callback);
+    });
+
+  };
+
+})();